Schema updated - added insert time and link type
I would like to start recording matches that were considered "good
enough" for the tool to add an item to immich albums, not just what the
tool copied to immich itself.
For instance, the tool adds to albums items that matched with type
"MatchedUnique", i.e. filename and metadata matched. But we don't record
them into the local db.

This info might be useful later on if we want to take care of deletions.

Also added insert times for the 2 tables that did not have it already.