Rewrote the main logic. Still WIP.
Broken down the whole process into three stages:

1. Scan
2. Search
3. Write

The first two stages are read-only, last commits the changes.

Scan produces a list of items that will be processed. It is just a list
of media items and albums from gphotos.

This steps attempts to associate gphoto elements with those from immich.
As I wanted to keep this step read-only (for immich+db) an association
could be "unresolved" and will be handled in the following stage.

Writes results to immich and the local db. This could be copying media
items from google to immich, creating immich albums, adding items to
immich albums and updating the local db. Write stage checks for the
read-only flag.

Hm, I completely skip the write stage if read-only is set. TODO

- concurrency, especially for item copying
- early exit (exit after encountering the first album without new items)
- instead of checking read-only inside of write() a bunch of times, just
  call a different method to print the actions.
- progress bars